rename : 상태 관리 폴더 이름 변경
