+++ client/src/App.tsx
@@ -4,12 +4,13 @@ import MainPage from '@/pages/MainPage/MainPage'
 import BroadcastPage from '@/pages/BroadcastPage/BroadcastPage'
 import { ThemeProvider } from 'styled-components'
 import { useRecoilValue } from 'recoil'
-import { themeState, ThemeFlag } from '@/state/theme'
+import { themeState, ThemeFlag } from '@/states/theme'
 import { lightTheme, darkTheme } from '@/styles/theme'
 
 function App() {
   const currentTheme = useRecoilValue(themeState)
   const theme = currentTheme === ThemeFlag.light ? lightTheme : darkTheme
+
   return (
     <ThemeProvider theme={theme}>
       <BrowserRouter>
